requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiomysql==0.2.0
aiosignal==1.3.1
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmyui==1.9.3
cryptography==41.0.7
cybersource-rest-client-python==0.0.49
DateTime==5.3
Deprecated==1.2.14
dicttoxml==1.7.16
discord-webhook==1.3.1
distlib==0.3.8
filelock==3.15.4
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.0
lxml==4.9.3
markdown2==2.4.12
MarkupSafe==2.1.5
multidict==6.0.5
munch==4.0.0
mysql-connector==2.2.9
mysql-connector-python==8.3.0
numpy==2.0.0
orjson==3.9.13
packaging==24.1
pandas==2.2.2
pillow==10.2.0
pkgconfig==1.5.5
platformdirs==4.2.2
priority==2.0.0
psutil==6.0.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==23.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
Quart==0.19.4
redis==5.0.0
requests==2.31.0
setuptools==69.0.2
six==1.16.0
soupsieve==2.5
timeago==1.0.16
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.26.3
virtualfish==2.5.9
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
zope.interface==6.1
